Cleanup and fix memory leak
